feat : 재고 관련 수정 및 매출 관련 기능 개발

+++ src/main/java/com/nagane/franchise/order/dao/OrderRepository.java
@@ -22,7 +22,7 @@ public interface OrderRepository extends JpaRepository<Order, Long> {
 	List<Order> findByStoreNoAndState(@Param("storeNo") Long storeNo);
 
 	/* 조회 당일 날짜와 같은 날 들어온 주문 개수를 세는 쿼리(점포 번호 기준) */
-	Long countByStore_StoreNoAndOrderDateBetween(Long storeNo, LocalDateTime startOfDay, LocalDateTime endOfDay);
+	Integer countByStore_StoreNoAndOrderDateBetween(Long storeNo, LocalDateTime startOfDay, LocalDateTime endOfDay);
 
 	/* 결제 일시와 상태(1)와 가맹점 번호로 레코드 리스트 반환 */
 	@Query("SELECT o FROM Order o WHERE o.state = 1 AND o.store.storeNo = :storeNo AND o.orderDate BETWEEN :startDate AND :endDate")
